Rename folders to repos on the back-end
