+++ server/graphql/resolvers/index.js
@@ -3,10 +3,10 @@ const activityResolver = require('./activity');
 const attemptingResolver = require('./attempting');
 const commentResolver = require('./comments');
 const issueResolver = require('./issues');
-const organizationResolver = require('./organizations');
 const paymentResolver = require('./payments');
 const pullRequestResolver = require('./pullRequests');
 const questionResolver = require('./questions');
+const repoResolver = require('./repos');
 const statsResolver = require('./stats');
 const userResolver = require('./users');
 const watchingResolver = require('./watching');
@@ -18,10 +18,10 @@ const rootResolver = {
   ...commentResolver,
   ...issueResolver,
   ...objectScalerType,
-  ...organizationResolver,
   ...paymentResolver,
   ...pullRequestResolver,
   ...questionResolver,
+  ...repoResolver,
   ...statsResolver,
   ...userResolver,
   ...watchingResolver,
